Load of changes, bunch of litmas changes and some bugfixes

+++ Shaders/LitMAS/LitMASInclude/ShaderInjector/Injections/Injection_DetailMap_CBuffer.hlsl
@@ -1,4 +1,8 @@
-//#!INJECT_BEGIN MATERIAL_CBUFFER 0
+//#!INJECT_BEGIN MATERIAL_CBUFFER_FLOAT_VECTORS 0
     float4 _DetailMap_ST;
-    float  _Details;
+    //half4  _DetailScale;
+//#!INJECT_END
+
+//#!INJECT_BEGIN MATERIAL_CBUFFER_HALF_SCALARS 0
+    half  _Details;
 //#!INJECT_END
